Ajustes para arreglar lo que hizo comentar con ia, arregaldos controladores, arregladas rutas, arreglados los modelos, optimizacion de scripts de js,se modifico el poperties apra uso de variables locales, se agrego el logo del servidor a las apginas thymeleaf.Se arreglo el fallo de redireccion de logout. Retoques menores en el front. Agregado servicio (Faltante) para hgestion de mail, no funcional,Termiandas pruebas del controlador. Porceder a empaquetado, pruebas , creacionde exe y documentacion.

+++ ServidorServicioAPI/GestorAPI/src/main/java/Proyecto/GestorAPI/controllers/UserController.java
@@ -91,7 +91,8 @@ public class UserController {
             String newUrl = storageService.saveImageData(STORAGE_BASE_PATH, file);
 
             // Eliminar imagen anterior si existe
-            if (oldUrl != null && !oldUrl.isEmpty()) {
+
+            if (oldUrl != null && !oldUrl.isEmpty() && !oldUrl.startsWith("hhtp")) {
                 storageService.deleteImageData(oldUrl);
             }
